removed topLevel benchmark
