+++ package.json
@@ -9,7 +9,6 @@
   "scripts": {
     "dev": "pnpm -r dev",
     "build": "pnpm -r build",
-    "bench": "node scripts/index.bench.mjs",
     "eslint": "eslint packages/**/*.{ts,js,vue}",
     "stylelint": "stylelint packages/**/*.{vue,css}",
     "lint": "pnpm eslint && pnpm stylelint",
@@ -22,7 +21,6 @@
     "bumpp": "^8.2.1",
     "eslint": "^8.33.0",
     "eslint-plugin-youcan": "workspace:*",
-    "glob": "^10.3.10",
     "mitata": "^0.1.6",
     "postcss-html": "^1.5.0",
     "rimraf": "^3.0.2",
@@ -31,7 +29,6 @@
     "stylelint-config-standard-scss": "^6.1.0",
     "stylelint-config-standard-vue": "^1.0.0",
     "taze": "^0.8.5",
-    "ts-node": "^10.9.2",
     "typescript": "^4.9.5"
   }
 }
