--- scripts/test.bench.mjs
@@ -1,19 +1,0 @@
-import { bench, run } from 'mitata';
-
-const fn = () => {};
-
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-bench('noop', fn);
-
-await run({ percentiles: false });
